[fix: Quitar errores de la consola y añadir comprobación de stock para los picos]

+++ backend/product/signals.py
@@ -30,7 +30,7 @@ def auto_delete_file_on_change(sender, instance, **kwargs):
 
     new_file = instance.image
     if not old_file == new_file:
-        if os.path.isfile(old_file.path):
+        if old_file and os.path.isfile(old_file.path):
             os.remove(old_file.path)
             print("Old image file deleted")
         else:
